% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iclogcondist_functions.R
\name{find_qsi}
\alias{find_qsi}
\title{Compute qsi Matrix}
\usage{
find_qsi(is, tau_no_Inf)
}
\arguments{
\item{is}{Indices of nodes.}

\item{tau_no_Inf}{Unique time points excluding infinity.}
}
\value{
qsi matrix.
}
\description{
This function computes the qsi matrix for specified indices and time points.
}
