% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iclogcondist_functions.R
\name{get_F_at_x}
\alias{get_F_at_x}
\title{Generic Function to compute F at X}
\usage{
get_F_at_x(object, ...)
}
\arguments{
\item{object}{An object for which the method is defined.}

\item{...}{Additional arguments passed to the method.}
}
\description{
This is a generic function to compute \eqn{F(x)} for object class \code{iclogcondist}. For usage details, please refer to function \code{get_F_at_x.iclogcondist}
}
